+++ src/main/resolveHTMLPathUtil.ts
